requirements.txt
appdirs==1.4.3
attrs==19.1.0
black==19.3b0
Click==7.0
numpy==1.17.0
opencv-python==4.1.0.25
Pillow==6.1.0
psycopg2==2.8.3
PyQt5==5.13.0
PyQt5-sip==4.19.18
scipy==1.3.1
six==1.12.0
structlog==19.1.0
toml==0.10.0
matplotlib==2.1.2
